refactor: migrate from elements to compounds and add support for parentheses
Compounds (collections of elements) are now used in place in of elements, allowing for support of reading parentheses in expressions. Formulas are also parsed to compounds which leverages the existing parser for parsing to elements.

BREAKING CHANGE: Since compounds are now used in favor of elements, many changes have taken place.